Remove config files
